+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/di/AppModule.kt
@@ -34,8 +34,7 @@ object AppModule {
 	@Singleton
 	@ApplicationScope
 	@Provides
-	fun providesApplicationScope(@DefaultDispatcher defaultDispatcher: CoroutineDispatcher): CoroutineScope =
-		CoroutineScope(SupervisorJob() + defaultDispatcher)
+	fun providesApplicationScope(@DefaultDispatcher defaultDispatcher: CoroutineDispatcher): CoroutineScope = CoroutineScope(SupervisorJob() + defaultDispatcher)
 
 	@Singleton
 	@Provides