+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/MainActivity.kt
@@ -74,6 +74,7 @@ import net.nymtech.nymvpn.ui.screens.settings.logs.LogsScreen
 import net.nymtech.nymvpn.ui.screens.settings.privacy.PrivacyScreen
 import net.nymtech.nymvpn.ui.screens.settings.support.SupportScreen
 import net.nymtech.nymvpn.ui.screens.splash.SplashScreen
+import net.nymtech.nymvpn.ui.screens.settings.tunneling.SplitTunnelingScreen
 import net.nymtech.nymvpn.ui.screens.welcome.WelcomeScreen
 import net.nymtech.nymvpn.ui.theme.NymVPNTheme
 import net.nymtech.nymvpn.ui.theme.Theme
@@ -123,6 +124,7 @@ class MainActivity : AppCompatActivity() {
 			var navHeight by remember { mutableStateOf(0.dp) }
 			val density = LocalDensity.current
 			var hideBackButtonInNavBar by remember { mutableStateOf(false) }
+			var onBackClickEventFromRoute by remember { mutableStateOf<Route?>(null) }
 
 			LaunchedEffect(configurationChange) {
 				if (configurationChange) {
@@ -177,6 +179,9 @@ class MainActivity : AppCompatActivity() {
 										}
 									},
 									hideBackButton = hideBackButtonInNavBar,
+									onBackClick = {
+										onBackClickEventFromRoute = it
+									},
 								)
 							},
 							snackbarHost = {
@@ -301,6 +306,15 @@ class MainActivity : AppCompatActivity() {
 								}
 								composable<Route.Account> {
 									AccountInfoScreen(appState)
+								}
+								composable<Route.SplitTunneling> {
+									SplitTunnelingScreen(
+										appState = appState,
+										onBackEventConsume = {
+											onBackClickEventFromRoute = null
+										},
+										onBackClickEventTriggered = onBackClickEventFromRoute == Route.SplitTunneling,
+									)
 								}
 							}
 						}