+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/screens/settings/components/LegalSection.kt
@@ -1,5 +1,6 @@
 package net.nymtech.nymvpn.ui.screens.settings.components
 
+import android.content.res.Configuration
 import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.outlined.ArrowRight
@@ -9,15 +10,15 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
 import net.nymtech.nymvpn.R
-import net.nymtech.nymvpn.ui.Route
 import net.nymtech.nymvpn.ui.common.buttons.surface.SelectionItem
-import net.nymtech.nymvpn.ui.common.navigation.LocalNavController
+import net.nymtech.nymvpn.ui.theme.NymVPNTheme
+import net.nymtech.nymvpn.ui.theme.Theme
 import net.nymtech.nymvpn.ui.theme.iconSize
 
 @Composable
-fun LegalSection() {
-	val navController = LocalNavController.current
+fun LegalSection(onLegalClick: () -> Unit) {
 	SettingsGroup(
 		items = listOf(
 			SelectionItem(
@@ -26,6 +27,7 @@ fun LegalSection() {
 						Icons.AutoMirrored.Outlined.ArrowRight,
 						stringResource(R.string.go),
 						modifier = Modifier.size(iconSize),
+						tint = MaterialTheme.colorScheme.onBackground,
 					)
 				},
 				title = {
@@ -34,8 +36,16 @@ fun LegalSection() {
 						style = MaterialTheme.typography.bodyLarge.copy(MaterialTheme.colorScheme.onSurface),
 					)
 				},
-				onClick = { navController.navigate(Route.Legal) },
+				onClick = onLegalClick,
 			),
 		),
 	)
+}
+
+@Composable
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
+internal fun PreviewLegalSection() {
+	NymVPNTheme(Theme.default()) {
+		LegalSection {}
+	}
 }