+++ nym-vpn-android/CHANGELOG.md
@@ -7,17 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.5.0]
+
 ### Added
 - Node details added in quick settings notification (https://github.com/nymtech/nym-vpn-client/pull/4029)
 - AmneziaWG section added to censorship screen (https://github.com/nymtech/nym-vpn-client/pull/4033)
 - Revoke implementation added for VPN service (https://github.com/nymtech/nym-vpn-client/pull/4041)
+- Account screen added (https://github.com/nymtech/nym-vpn-client/pull/4060)
 
+- Android: Bi-mode Split Tunneling (https://github.com/nymtech/nym-vpn-client/pull/4049
+-
 ### Changed
+- UI changes and updates for Settings screen (https://github.com/nymtech/nym-vpn-client/pull/4060)
+- Fixes for Privacy & Censorship screens (https://github.com/nymtech/nym-vpn-client/pull/4060)
+- UI update for Navigation bar style (https://github.com/nymtech/nym-vpn-client/pull/4060)
 
 ### Fixed
 - Connection error notification cleared after successfully connecting (https://github.com/nymtech/nym-vpn-client/pull/4029)
 - Fix connection drops after language change (https://github.com/nymtech/nym-vpn-client/pull/4034)
 - Fix QUIC autostart issue (https://github.com/nymtech/nym-vpn-client/pull/4040)
+- Fix for Sentry toggle (https://github.com/nymtech/nym-vpn-client/pull/4060)
 
 ## [2.4.0]